#100: moved Logger
