+++ src/main/scala/dbtarzan/gui/Database.scala
@@ -7,13 +7,8 @@ import dbtarzan.gui.foreignkeys.{VirtualForeignKeysEditor, VirtualForeignKeysEdi
 import dbtarzan.gui.interfaces.TControlBuilder
 import dbtarzan.gui.util.{FilterText, JFXUtil}
 import dbtarzan.localization.Localization
+import dbtarzan.log.actor.Logger
 import dbtarzan.messages.*
-import scalafx.Includes.*
-import scalafx.event.ActionEvent
-import scalafx.scene.Parent
-import scalafx.scene.control.*
-import scalafx.scene.layout.{BorderPane, FlowPane, VBox}
-import scalafx.stage.Stage
 
 /* A panel containing all the tabs related to a database */
 class Database (dbActor : ActorRef, guiActor : ActorRef, databaseId : DatabaseId, localization : Localization, tableIds: List[TableId], log: Logger) extends TControlBuilder {
