+++ src/main/scala/dbtarzan/gui/Table.scala
@@ -5,14 +5,9 @@ import dbtarzan.gui.interfaces.TControlBuilder
 import dbtarzan.gui.table.*
 import dbtarzan.gui.util.JFXUtil
 import dbtarzan.localization.Localization
+import dbtarzan.log.actor.Logger
 import dbtarzan.messages.*
 import org.apache.pekko.actor.ActorRef
-import scalafx.Includes.*
-import scalafx.collections.ObservableBuffer
-import scalafx.scene.Parent
-import scalafx.scene.control.TableColumn.*
-import scalafx.scene.control.{SelectionMode, TableColumn, TableView}
-import scalafx.scene.image.ImageView
 
 
 /** The GUI table control showing the content of a database table in a GUI table*/
