+++ src/main/scala/dbtarzan/gui/rowdetails/RowDetailsViewFields.scala
@@ -4,7 +4,7 @@ import dbtarzan.db.{DBTable, Field, Row}
 import dbtarzan.gui.interfaces.TControlBuilder
 import dbtarzan.gui.util.JFXUtil
 import dbtarzan.localization.Localization
-import dbtarzan.messages.Logger
+import dbtarzan.log.actor.Logger
 import scalafx.geometry.{Insets, Pos}
 import scalafx.scene.Parent
 import scalafx.scene.control.ScrollPane
